👌 IMPROVE: Move constants to file
