+++ frontend/app/(guest)/register/constants.ts
@@ -1,0 +1,8 @@
+export const ALLOWED_UPLOADED_FILE_TYPES = [
+  "image/png",
+  "image/jpg",
+  "image/jpeg",
+  "application/pdf",
+];
+
+export const MAX_UPLOADED_FILE_SIZE = 2000000;
